data/scripts/catalogue/requirements.txt
certifi==2024.7.4
charset-normalizer==2.1.1
idna==3.7
numpy==1.26.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
requests==2.32.3
setuptools==78.1.1
six==1.16.0
molgenis_emx2_pyclient==11.61.2
urllib3==2.5.0

